refactor(Line): 한 줄을 만드는 메서드를 Line의 정적 팩토리 메서드로 추가
